+++ packages/lumx-core/src/js/components/InputHelper/Stories.ts
@@ -1,10 +1,13 @@
 import { Kind } from '@lumx/core/js/constants';
 import { getSelectArgType } from '@lumx/core/stories/controls/selectArgType';
 
+import { InputHelper } from '.';
+
 export const Default = {
     argTypes: {
         kind: getSelectArgType(Kind),
     },
+    args: { ...InputHelper.defaultProps },
 };
 
 /**